线程处理消息

+++ MPCVer2.0.0/src/Pressordelegate/downloadmanager.h
@@ -11,7 +11,7 @@ public:
     explicit downloadmanager(QObject *parent = 0);
        virtual void run() ;
 signals:
-	  void sendValue(int value);
+	  void sendValue(double value);
     
 public slots:
